+++ providers/capa/private/capa_suite_test.go
@@ -7,7 +7,8 @@ import (
 	. "github.com/onsi/gomega"
 
 	"github.com/giantswarm/cluster-test-suites/internal/suite"
-	"github.com/giantswarm/cluster-test-suites/providers/capa"
+
+	"github.com/giantswarm/cluster-standup-teardown/pkg/clusterbuilder/providers/capa"
 )
 
 const KubeContext = "capa-private-proxy"